Fixed the containment handling of the EventAdapter
The EventAdapter used to overwrite the add but not the remove method of
the EContentAdapter and thus broke its protocol. The adapter prevented
itself from being re-added to a notifier if the notifier was 'logically'
contained within a resource and 'physically' contained via a containment
reference.

This fix makes sure that adapter remains attached to a notifier until
both of these containments are unset.